requirements.txt

streamlit>=1.38.0
python-docx>=1.1.2
beautifulsoup4>=4.12.2
lxml>=5.3.0
mammoth>=1.6.0
pdfplumber>=0.11.0
